Added patch endpoint

+++ backend/routes/todos.js
@@ -47,7 +47,7 @@ router.post("/", async(req, res) => {
         console.log(error.message)
     } 
 })
-
+//Edit the body of a document
 router.put("/:id", async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
@@ -83,7 +83,26 @@ router.put("/:id", async (req, res) => {
     }
 })
 
+router.patch("/:id", async (req, res) => {
+    const todo = await Todo.findById(req.params.id)
+
+    if(!todo) return res.status(404).send("Todo not found...")
+
+    try {
+    const updatedTodo = await Todo.findByIdAndUpdate(req.params.id, {
+        isComplete: !todo.isComplete
+    })
+        res.send(updatedTodo)
+    } catch (error) {
+        res.status(500).send(error.message)
+        console.log(error.message)
+    }
+})
+
 router.delete("/:id", async (req, res) => {
+    const todo = await Todo.findById(req.params.id)
+
+    if(!todo) return res.status(404).send("Todo not found...")
 
     try {
         const deletedTodo = await Todo.findByIdAndDelete(req.params.id)
